Change project structure
I don't really like having content directly in the top directory, so
we moved it to a `site` folder, somewhat symmetric to the `_site`
folder where things are generated.  This means wee also needed to
rename the site generator name, which is fine -- I think the new name
is better.